requirements.txt
fastapi
uvicorn[standard]
google-cloud-storage
pillow
python-multipart
moondream
requests
python-dotenv 